finish mutinode test for mpi and utc.
Add cpu bind "AffinityConfig.h"  related things. Different bind options. When creating a task, last argument can be set to choose the disired bind style.

+++ utc-0.1/include/AffinityUtilities.h
@@ -101,7 +101,7 @@ inline void setAffinityLinux(std::vector<int> cpus){
 	std::vector<int> ret;
 	CPU_ZERO(&cpuset);
 	thread = pthread_self();
-	for(int i=0; i<cpus.size(); i++)
+	for(int i=0; i<(int)cpus.size(); i++)
 		CPU_SET(cpus[i], &cpuset);
 	int s;
 	s = pthread_setaffinity_np(thread, sizeof(cpu_set_t), &cpuset);
